splitting SaveToHadoop from SaveToHadoopMethods
matching SaveToS3 layout
